requirements.txt
# Core Agent Framework
strands-agents>=0.1.0

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# Frontend
streamlit>=1.31.0
streamlit-audio-recorder>=0.0.8

# Image Processing
pillow>=10.2.0

# Environment Management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
hypothesis>=6.98.0
pytest-asyncio>=0.23.0

# Utilities
requests>=2.31.0
